requirements.txt
absl-py==0.13.0
alabaster==0.7.12
anipose==0.9.0
aniposelib==0.4.1
anyio==3.3.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bidict==0.21.2
bleach==3.3.1
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
colorcet==2.0.6
contextvars==2.4
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
datajoint==0.13.2
decorator==5.0.9
deeplabcut==2.1.8.2
deeplabcut-live==0.0.3
defusedxml==0.7.1
dnspython==1.16.0
docutils==0.17.1
easydict==1.9
entrypoints==0.3
eventlet==0.30.1
execnet==1.9.0
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-SocketIO==5.1.1
future==0.18.2
gast==0.5.2
greenlet==1.0.0
grpcio==1.39.0
h5py==2.10.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
imagesize==1.2.0
imgaug==0.4.0
immutables==0.16
importlib-metadata==4.6.3
iniconfig==1.1.1
intel-openmp==2021.3.0
ipykernel==5.3.4
ipympl==0.7.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
llvmlite==0.33.0
Markdown==3.3.4
MarkupSafe==2.0.1
mat4py==0.4.3
matlabengineforpython==R2020b
matplotlib==3.3.4
minio==7.1.0
mistune==0.8.4
mock==4.0.3
moviepy==1.0.1
mpmath==1.2.1
msgpack==1.0.2
msgpack-numpy==0.4.7.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nidaqmx==0.5.7
notebook==6.4.0
npTDMS==1.1.0
numba==0.50.1
numexpr==2.7.3
numpy==1.19.5
olefile==0.46
opencv-contrib-python==4.5.3.56
otumat==0.2.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
param==1.11.1
parso==0.8.2
pathlib2==2.3.6
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.3.2
pip==21.1.3
pkginfo==1.7.1
pluggy==0.13.1
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
py-cpuinfo==5.0.0
PyAudio==0.2.11
pycparser==2.20
pyct==0.4.8
pydot==1.4.2
Pygments==2.9.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
pytest-forked==1.3.0
pytest-timeout==1.4.2
pytest-xdist==2.3.0
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.1.0
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.5.4
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spinnaker-python==2.0.0.147
statsmodels==0.12.2
sympy==1.7.1
tables==3.6.1
tabulate==0.8.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorpack==0.11
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2020.9.3
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
twine==3.4.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wxPython==4.0.5
zipp==3.5.0
